Fix bugs in reading properties
Property keys and values are strings, but were initially being read as
ints. The property reading loop in readNode was stuck in an infinite
while loop because propertyID had not been properly set in the
constructor. Property value was also longer than the 4 bytes allowed,
which was overwriting the nextPropID.