Change mode used to open files for writing to r+b
The mode used to open Node, Property, Relationship, and Label store
files for writing was previously "ab"; this mode makes it so that
these store files were opened so that any writes would be appended
to the end of the files, disregarding any calls to seek(). Change
mode used to open store files for writing to "r+b" so that files
will be opened for reading and writing in binary.