Fix bug with adding properties when reading Nodes
Previously, when reading Nodes, the properties that were read in
would be added to the Node before the ID of the next property was
read, causing the next property ID of all properties to be -1 (i.e.
that there is no next property). Change reading nodes so that
properties are added after the ID of the next property is read in.