Implement simple expression evaluation
We skipped all of the functions that have to do with string literals,
variable evaluation, etc. However, now if you do something like
CREATE (a {math: 1 + 1 - 1 * 3}), it will do the math correctly.
Next up is actually storing this data, and probably adding all the
string stuff we did not do.