Implement handling of relationships
Relationships, nodes, etc. that are supposed to be created from
create-statements are now properly handled, I guess. It's hard to
say what "properly" means.

It's assumed that relationships, if they're being created, connect
exactly two nodes (longer chains are possible in real Cypher). Also,
it (alongside everything else in the program) ignores variable
bindings. That'll have to be fixed.