Test Queries with End Node Type
Write test for queries with a specific end node type, such as “friends
of harry potter who own a cat”. Add third item to condition tuples for
BFS, an integer indicating which node is wanted at a given level. For
example, the second level of relationships is looking at an owner and
its cat, but do we want the owner or the cat in the result. 0 indicates
the higher node, 1 indicates the lower node, based on closeness to the
start node.