Add (untested!) replacement BFS code
The code fleshes out many of the structures and algorithms required
to get BFS working with the Dummy* types. What needs to be done is,
among other things, the abstraction of a helper function to complete
the implementation. Not to mention the linear scan through the nodes
and/or relationships!