Continue evaluation of RETURN statements
We've stubbed out a majority of the work which has yet to come,
but we did this because we've hit a crossroads: we apparently need
to copy-and-paste, or otherwise rewrite, a lot of the expression
evaluation code from the ANTLR section. We would like to not have
to do this, so we are waiting on further instruction/suggestions.

What was done? We are now dealing with the things returned by
the parser w.r.t. MATCH and RETURN, and should ostensibly be
pretty close to done with that stuff soon. We just need to
swap in "16" for "m.age" in all the relevant places, then re-do
calculator parsing; for example "m.age * 2" will become "16 * 2",
which will be parsed+evaluated to become "32".