Rename basicTest in /storage
This was causing some major errors when trying to merge branches
together.  I think this is a remnant from when we didn't have a
convention set up and we had a /storage directory and a /Storage
directory.  The next couple of commits and merges will try to
rectify this situation.
The error was: The following untracted working tree files would
be overwritten by checkout.