Add ability to clear data from prompt
To clear out old data, you had to delete the store files.
Now, 'clean' has been added to the REPL and will remove
all the files in /datafiles and remake them.
Currently, it does not restart the counter for the files,
but that shouldn't affect functionality and can be fixed
later.