Fix Write Label
Writing a label page to disk is just writing all labels in that page to
disk. When label storage manager is called to write label, the new
label is replaced in label page data in memory, but not written to
disk. Full page will be written to disk when buffer manager evicts page.